+++ tests/poetry.rs
@@ -2183,12 +2183,6 @@ fn test_build_backend_uv_errors() {
     error: - "packages_glob_from_to_sdist_wheel/**/*.py" from "poetry.packages.include" cannot be converted to uv, as it is configured to be added to both source distribution and wheels, and uses globs, which cannot be expressed with uv.
     error: - "text_file_sdist_wheel.txt" from "poetry.packages.include" cannot be converted to uv, as it is configured to be added to both source distribution and wheels, and is a file, which cannot be expressed with uv.
     error: - "text_file_wheel.txt" from "poetry.packages.include" cannot be converted to uv, as it is configured to be added to wheels only, and is a file, which cannot be expressed with uv.
-    error: - "packages_without_init" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
-    error: - "packages_without_init_root" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
-    error: - "packages_from_without_init" from "poetry.packages.include" cannot be converted to uv, as it uses "from", which cannot be expressed with uv.
-    error: - "packages_from_without_init" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
-    error: - "packages_from_without_init_root" from "poetry.packages.include" cannot be converted to uv, as it uses "from", which cannot be expressed with uv.
-    error: - "packages_from_without_init_root" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
     error: - "include_sdist_wheel" from "poetry.include" cannot be converted to uv, as it is configured to be added to both source distribution and wheels, which cannot be expressed with uv.
     error: - "include_wheel" from "poetry.include" cannot be converted to uv, as it is configured to be added to wheels only, which cannot be expressed with uv.
     error: - "include_wheel_2" from "poetry.include" cannot be converted to uv, as it is configured to be added to wheels only, which cannot be expressed with uv.
@@ -2225,12 +2219,6 @@ fn test_build_backend_uv_errors_dry_run() {
     error: - "packages_glob_from_to_sdist_wheel/**/*.py" from "poetry.packages.include" cannot be converted to uv, as it is configured to be added to both source distribution and wheels, and uses globs, which cannot be expressed with uv.
     error: - "text_file_sdist_wheel.txt" from "poetry.packages.include" cannot be converted to uv, as it is configured to be added to both source distribution and wheels, and is a file, which cannot be expressed with uv.
     error: - "text_file_wheel.txt" from "poetry.packages.include" cannot be converted to uv, as it is configured to be added to wheels only, and is a file, which cannot be expressed with uv.
-    error: - "packages_without_init" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
-    error: - "packages_without_init_root" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
-    error: - "packages_from_without_init" from "poetry.packages.include" cannot be converted to uv, as it uses "from", which cannot be expressed with uv.
-    error: - "packages_from_without_init" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
-    error: - "packages_from_without_init_root" from "poetry.packages.include" cannot be converted to uv, as it uses "from", which cannot be expressed with uv.
-    error: - "packages_from_without_init_root" from "poetry.packages.include" cannot be converted to uv, as it does not contain an "__init__.py" file, which is required by uv for packages.
     error: - "include_sdist_wheel" from "poetry.include" cannot be converted to uv, as it is configured to be added to both source distribution and wheels, which cannot be expressed with uv.
     error: - "include_wheel" from "poetry.include" cannot be converted to uv, as it is configured to be added to wheels only, which cannot be expressed with uv.
     error: - "include_wheel_2" from "poetry.include" cannot be converted to uv, as it is configured to be added to wheels only, which cannot be expressed with uv.
@@ -2462,6 +2450,86 @@ fn test_build_backend_implicit_package_src_and_no_src() {
 
     [tool.uv.build-backend]
     module-root = ""
+    source-include = ["text_file_sdist.txt"]
+    "#);
+
+    Command::new("uv")
+        .arg("build")
+        .current_dir(project_path)
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()
+        .unwrap();
+
+    let sdist_files_after = get_tar_gz_entries(&project_path.join("dist"), "foobar-0.1.0.tar.gz");
+    let wheel_files_after =
+        get_zip_entries(&project_path.join("dist"), "foobar-0.1.0-py3-none-any.whl");
+
+    assert_eq!(sdist_files_before, sdist_files_after);
+    assert_eq!(wheel_files_before, wheel_files_after);
+}
+
+#[test]
+fn test_build_backend_uses_namespace() {
+    let fixture_path = Path::new(FIXTURES_PATH).join("build_backend/uses_namespace");
+
+    let tmp_dir = tempdir().unwrap();
+    let project_path = tmp_dir.path();
+
+    copy_dir(&fixture_path, project_path).unwrap();
+
+    Command::new("uvx")
+        .arg("poetry")
+        .arg("build")
+        .current_dir(project_path)
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()
+        .unwrap();
+
+    let sdist_files_before = get_tar_gz_entries(&project_path.join("dist"), "foobar-0.1.0.tar.gz");
+    let wheel_files_before =
+        get_zip_entries(&project_path.join("dist"), "foobar-0.1.0-py3-none-any.whl");
+
+    remove_dir_all(project_path.join("dist")).unwrap();
+
+    assert_cmd_snapshot!(cli().arg(project_path).arg("--skip-lock"), @r"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+
+    ----- stderr -----
+    Successfully migrated project from Poetry to uv!
+
+    warning: The following warnings occurred during the migration:
+    warning: - Build backend was migrated to uv. It is highly recommended to manually check that files included in the source distribution and wheels are the same than before the migration.
+    ");
+
+    apply_filters!();
+    insta::assert_snapshot!(fs::read_to_string(project_path.join("pyproject.toml")).unwrap(), @r#"
+    [build-system]
+    requires = ["uv_build>=[LOWER_BOUND],<[UPPER_BOUND]"]
+    build-backend = "uv_build"
+
+    [project]
+    name = "foobar"
+    version = "0.1.0"
+    description = "A fabulous project."
+    authors = [{ name = "John Doe", email = "john.doe@example.com" }]
+    requires-python = ">=3.10"
+    classifiers = [
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
+        "Programming Language :: Python :: 3.13",
+        "Programming Language :: Python :: 3.14",
+    ]
+
+    [tool.uv.build-backend]
+    module-name = ["namespace"]
+    module-root = ""
+    namespace = true
     source-include = ["text_file_sdist.txt"]
     "#);